sync from flutter/engine

+++ lib/web_ui/dev/utils.dart
@@ -27,8 +27,8 @@ class FilePath {
       path.relative(_absolutePath, from: environment.webUiRootDir.path);
 
   @override
-  bool operator ==(dynamic other) {
-    return other is FilePath && _absolutePath == other._absolutePath;
+  bool operator ==(Object other) {
+    return other is FilePath && other._absolutePath == _absolutePath;
   }
 
   @override
@@ -67,7 +67,7 @@ Future<int> runProcess(
 }
 
 /// Runs [executable]. Do not follow the exit code or the output.
-void startProcess(
+Future<void> startProcess(
   String executable,
   List<String> arguments, {
   String workingDirectory,
@@ -215,7 +215,9 @@ void cleanup() async {
   // Delete temporary directories.
   if (temporaryDirectories.length > 0) {
     for (io.Directory directory in temporaryDirectories) {
-      directory.deleteSync(recursive: true);
+      if (!directory.existsSync()) {
+        directory.deleteSync(recursive: true);
+      }
     }
   }
